requirements.txt
#python==3.9


absl-py                      2.1.0
anyio                        4.6.2
argon2-cffi                  21.3.0
argon2-cffi-bindings         21.2.0
asttokens                    2.0.5
astunparse                   1.6.3
async-lru                    2.0.4
attrs                        24.2.0
Babel                        2.11.0
backcall                     0.2.0
beautifulsoup4               4.12.3
bleach                       6.2.0
blinker                      1.9.0
Brotli                       1.0.9
cachetools                   5.5.0
certifi                      2024.8.30
cffi                         1.17.1
charset-normalizer           3.3.2
click                        8.1.7
colorama                     0.4.6
comm                         0.2.1
contourpy                    1.3.0
cycler                       0.12.1
debugpy                      1.6.7
decorator                    5.1.1
defusedxml                   0.7.1
exceptiongroup               1.2.0
executing                    0.8.3
fastjsonschema               2.20.0
Flask                        3.1.0
flatbuffers                  24.3.25
fonttools                    4.55.0
gast                         0.4.0
glob2                        0.7
google-auth                  2.36.0
google-auth-oauthlib         1.0.0
google-pasta                 0.2.0
grpcio                       1.68.1
h11                          0.14.0
h5py                         3.8.0
httpcore                     1.0.2
httpx                        0.27.0
idna                         3.10
importlib_metadata           8.5.0
importlib_resources          6.4.5
ipykernel                    6.29.5
ipython                      8.15.0
ipywidgets                   8.1.2
itsdangerous                 2.2.0
jedi                         0.19.1
Jinja2                       3.1.4
joblib                       1.4.2
json5                        0.9.25
jsonschema                   4.23.0
jsonschema-specifications    2023.7.1
jupyter                      1.0.0
jupyter_client               8.6.0
jupyter-console              6.6.3
jupyter_core                 5.7.2
jupyter-events               0.10.0
jupyter-lsp                  2.2.0
jupyter_server               2.14.1
jupyter_server_terminals     0.4.4
jupyterlab                   4.2.5
jupyterlab-pygments          0.1.2
jupyterlab_server            2.27.3
jupyterlab-widgets           3.0.10
keras                        2.13.1
kiwisolver                   1.4.7
libclang                     18.1.1
Markdown                     3.7
MarkupSafe                   2.1.3
matplotlib                   3.7.1
matplotlib-inline            0.1.6
mistune                      2.0.4
nbclient                     0.8.0
nbconvert                    7.16.4
nbformat                     5.10.4
nest-asyncio                 1.6.0
notebook                     7.2.2
notebook_shim                0.2.3
numpy                        1.23.5
oauthlib                     3.2.2
opencv-python                4.7.0.68
opt_einsum                   3.4.0
overrides                    7.4.0
packaging                    24.1
pandas                       1.5.3
pandocfilters                1.5.0
parso                        0.8.3
pickleshare                  0.7.5
Pillow                       9.2.0
pip                          24.2
platformdirs                 3.10.0
ply                          3.11
prometheus_client            0.21.0
prompt-toolkit               3.0.43
protobuf                     4.25.5
psutil                       5.9.0
pure-eval                    0.2.2
pyasn1                       0.6.1
pyasn1_modules               0.4.1
pycparser                    2.21
Pygments                     2.15.1
pyparsing                    3.2.0
PyQt5                        5.15.10
PyQt5-sip                    12.13.0
PySocks                      1.7.1
python-dateutil              2.9.0.post0
python-json-logger           2.0.7
pytz                         2024.1
pywin32                      305.1
pywinpty                     2.0.10
PyYAML                       6.0.2
pyzmq                        25.1.2
qtconsole                    5.6.0
QtPy                         2.4.1
referencing                  0.30.2
requests                     2.32.3
requests-oauthlib            2.0.0
rfc3339-validator            0.1.4
rfc3986-validator            0.1.1
rpds-py                      0.10.6
rsa                          4.9
scikit-learn                 1.1.2
scipy                        1.13.1
Send2Trash                   1.8.2
setuptools                   75.1.0
sip                          6.7.12
six                          1.16.0
sniffio                      1.3.0
soupsieve                    2.5
stack-data                   0.2.0
tensorboard                  2.13.0
tensorboard-data-server      0.7.2
tensorflow                   2.13.0
tensorflow-estimator         2.13.0
tensorflow-hub               0.13.0
tensorflow-intel             2.13.0
tensorflow-io-gcs-filesystem 0.31.0
termcolor                    2.5.0
terminado                    0.17.1
threadpoolctl                3.5.0
tinycss2                     1.2.1
tomli                        2.0.1
tornado                      6.4.1
traitlets                    5.14.3
typing_extensions            4.5.0
urllib3                      2.2.3
wcwidth                      0.2.5
webencodings                 0.5.1
websocket-client             1.8.0
Werkzeug                     3.1.3
wheel                        0.44.0
widgetsnbextension           4.0.10
win-inet-pton                1.1.0
wrapt                        1.17.0
zipp                         3.21.0